XIVY-5084 Use own image that contains app.zip to deploy
Instead of using the axonivy-engine image direct build an own image
axonivy-engine-cluster-test with the axonivy-engine as base and add all
the configuration and projects to deploy to that image.
In the docker-compose use then the new axonivy-engine-cluster-test image

Remove entrypoint-master.sh and entrypoint-slave.sh as this is now
handled correct by cli create-db command.
Remove unneeded configuration from ivy.yaml